+++ Extensions/Announcers/GithubAnnouncements/GithubAnnouncements.cs
@@ -26,10 +26,10 @@ namespace GithubAnnouncements
         const string LatestCommentsKey = "Comments";
         const string WatchersKey = "Watchers";
         const string ForkStatusKey = "ForkStatus";
+        const string Account = "Code52";
+        const string Repo = "jibbr";
 
         readonly ISettingsService _storage;
-        readonly string _account;
-        readonly string _repo;
         readonly string _apiUrl;
         readonly WebClient _client = new WebClient();
 
@@ -37,9 +37,7 @@ namespace GithubAnnouncements
         public GithubAnnouncements(ISettingsService storage)
         {
             _storage = storage;
-            _account = "Code52";
-            _repo = "jibbr";
-            _apiUrl = string.Format(UrlFormat, _account, _repo);
+            _apiUrl = string.Format(UrlFormat, Account, Repo);
         }
 
         public TimeSpan Interval
@@ -62,36 +60,16 @@ namespace GithubAnnouncements
             var latestCommit = commits.FirstOrDefault();
             if (latestCommit == null) return;
 
-            var lastCommit = "";
-            if (_storage.ContainsKey(LatestCommitKey))
-            {
-                lastCommit = _storage.Get<string>(LatestCommitKey);
-            }
-
+            var lastCommit = GetValue(LatestCommitKey, () => "");
             var latestSHA = latestCommit.commit.tree.sha.ToString();
 
             if (lastCommit == latestSHA)
                 return;
 
-            // set new value;
             _storage.Set(LatestCommitKey, latestSHA);
             _storage.Save();
 
-            // find new commits since 
-            var groupedCommits = commits.TakeWhile(c => c.commit.tree.sha != lastCommit).GroupBy(c => c.committer.login);
-
-            // create string to send as message
-            var sb = new StringBuilder();
-            sb.AppendFormat("github: there are new commits in the {0}/{1} repository\r\n", _account, _repo);
-            foreach (var committer in groupedCommits)
-            {
-                sb.AppendFormat("{0} had {1} commits\r\n", committer.Key, committer.Count());
-            }
-
-            // split into rows and send to rooms
-            var rows = sb.ToString().Split(new[] { "\r\n" }, StringSplitOptions.RemoveEmptyEntries);
-            foreach (var row in rows)
-                bot.SayToAllRooms(row);
+            bot.ProcessCommits(RepositoryName, lastCommit, commits);
         }
 
         private void NotifyPullRequests(Bot bot)
@@ -99,74 +77,13 @@ namespace GithubAnnouncements
             var openPullRequests = GetResponse<IEnumerable<dynamic>>(GetFullUrl(OpenPullRequestsFeed));
             var closedPullRequests = GetResponse<IEnumerable<dynamic>>(GetFullUrl(ClosedPullRequestsFeed));
 
-            IDictionary<int, string> existingPullRequests = new Dictionary<int, string>();
+            var existingPullRequests = GetValue<IDictionary<int, string>>(MergeRequestsKey, () => new Dictionary<int, string>());
 
-            if (_storage.ContainsKey(MergeRequestsKey))
-            {
-                existingPullRequests = _storage.Get<IDictionary<int, string>>(MergeRequestsKey);
-            }
-
-            ProcessClosedPullRequests(bot, closedPullRequests, existingPullRequests);
-
-            ProcessOpenPullRequests(bot, existingPullRequests, openPullRequests);
+            bot.ProcessClosedPullRequests(closedPullRequests, existingPullRequests);
+            bot.ProcessOpenPullRequests(RepositoryName, existingPullRequests, openPullRequests);
 
             _storage.Set(MergeRequestsKey, existingPullRequests);
             _storage.Save();
-        }
-
-        private static void ProcessClosedPullRequests(Bot bot, IEnumerable<dynamic> closedPullRequests, IDictionary<int, string> existingPullRequests)
-        {
-            foreach (var request in closedPullRequests)
-            {
-                int id;
-                if (!int.TryParse(request.number.ToString(), out id))
-                    continue;
-
-                if (!existingPullRequests.ContainsKey(id))
-                    continue;
-
-                // send message
-                string firstLine = string.Format("github: {0}'s pull request '{1}' has been closed",
-                                                 request.user.login,
-                                                 request.title);
-
-                var result = request.merged_at.ToString();
-                var secondLine = !string.IsNullOrWhiteSpace(result)
-                                     ? "The merge request was accepted. Awesome!"
-                                     : "The merge request was not accepted. Sadface!";
-
-                bot.SayToAllRooms(firstLine);
-                bot.SayToAllRooms(secondLine);
-
-                // cleanup request
-                existingPullRequests.Remove(id);
-            }
-        }
-
-        private void ProcessOpenPullRequests(Bot bot, IDictionary<int, string> existingPullRequests, IEnumerable<dynamic> openPullRequests)
-        {
-            foreach (var request in openPullRequests)
-            {
-                int id;
-                if (!int.TryParse(request.number.ToString(), out id))
-                    continue;
-
-                if (existingPullRequests.ContainsKey(id))
-                    continue;
-
-                // send message
-                string firstLine = string.Format("github: {0} has opened a pull request named '{1}' for {2}/{3}",
-                                                 request.user.login,
-                                                 request.title,
-                                                 _account,
-                                                 _repo);
-                string secondLine = string.Format("Please review the request at {0}", request.html_url);
-                bot.SayToAllRooms(firstLine);
-                bot.SayToAllRooms(secondLine);
-
-                // track request
-                existingPullRequests.Add(id, "open");
-            }
         }
 
         private void NotifyForks(Bot bot)
@@ -174,13 +91,8 @@ namespace GithubAnnouncements
             var forks = GetResponse<IEnumerable<dynamic>>(GetFullUrl(ProjectForksFeed));
 
             var feeds = forks.ToDictionary(f => f.owner.login, f => f.url);
-
-            IDictionary<string, string> existingForkStatus = new Dictionary<string, string>();
-            if (_storage.ContainsKey(ForkStatusKey))
-            {
-                existingForkStatus = _storage.Get<IDictionary<string, string>>(ForkStatusKey);
-            }
-
+            var existingForkStatus = GetValue<IDictionary<string, string>>(ForkStatusKey, () => new Dictionary<string, string>());
+            
             foreach (var fork in feeds)
             {
                 NotifyForkStatus(fork, existingForkStatus, bot);
@@ -192,31 +104,10 @@ namespace GithubAnnouncements
 
         private void NotifyForkStatus(KeyValuePair<dynamic, dynamic> fork, IDictionary<string, string> existingForkStatus, Bot bot)
         {
+            string id = fork.Key.ToString();
             string url = fork.Value.ToString() + "/commits";
             var commits = GetResponse<IEnumerable<dynamic>>(url).ToList();
-            var latestCommit = commits.FirstOrDefault();
-            if (latestCommit == null) return;
-
-            string id = fork.Key.ToString();
-
-            var latestSHA = latestCommit.commit.tree.sha.ToString();
-            if (!existingForkStatus.ContainsKey(id))
-            {
-                existingForkStatus.Add(id, latestSHA);
-                return;
-            }
-
-            var lastCommit = existingForkStatus[id];
-            if (lastCommit == latestSHA)
-                return;
-
-            existingForkStatus[id] = latestSHA;
-
-            // find new commits since 
-            var newCommits = commits.TakeWhile(c => c.commit.tree.sha != lastCommit);
-
-            // create string to send as message
-            bot.SayToAllRooms(string.Format("{0} added {1} new commits to his fork", id, newCommits.Count()));
+            bot.ProcessForkStatus(id, existingForkStatus, commits);
         }
 
         private void NotifyIssues(Bot bot)
@@ -232,8 +123,7 @@ namespace GithubAnnouncements
             }
 
             bot.ProcessClosedIssues(closedIssues, existingIssues);
-            var repo = string.Format("{0}/{1}", _account, _repo);
-            bot.ProcessOpenIssues(repo, openIssues, existingIssues);
+            bot.ProcessOpenIssues(RepositoryName, openIssues, existingIssues);
 
             _storage.Set(IssuesKey, existingIssues);
             _storage.Save();
@@ -241,31 +131,10 @@ namespace GithubAnnouncements
 
         private void NotifyWatchers(Bot bot)
         {
-            var watchers = GetResponse<IEnumerable<dynamic>>(GetFullUrl(ProjectWatchersFeed));
+            var currentWatchers = GetResponse<IEnumerable<dynamic>>(GetFullUrl(ProjectWatchersFeed));
+            var existingWatchers = GetValue<IList<string>>(WatchersKey, () => new List<string>());
 
-            var currentWatchers = watchers.Select(c => c.login.ToString()).Cast<string>().ToList();
-
-            IList<string> existingWatchers = new List<string>();
-            if (_storage.ContainsKey(WatchersKey))
-            {
-                existingWatchers = _storage.Get<IList<string>>(WatchersKey);
-            }
-
-            var newWatchers = currentWatchers.Except(existingWatchers).ToList();
-            foreach (var w in newWatchers)
-            {
-                var repoName = string.Format("{0}/{1}", _account, _repo);
-                bot.SayToAllRooms(string.Format("{0} is now watching {1}", w, repoName));
-            }
-
-            var noLongerWatching = existingWatchers.Except(currentWatchers).ToList();
-            foreach (var w in noLongerWatching)
-            {
-                var repoName = string.Format("{0}/{1}", _account, _repo);
-                bot.SayToAllRooms(string.Format("{0} is no longer watching {1}", w, repoName));
-            }
-
-            existingWatchers = existingWatchers.Concat(newWatchers).Except(noLongerWatching).ToList();
+            existingWatchers = bot.ProcessWatchers(RepositoryName, existingWatchers, currentWatchers);
 
             _storage.Set(WatchersKey, existingWatchers);
             _storage.Save();
@@ -280,6 +149,16 @@ namespace GithubAnnouncements
         {
             var response = _client.DownloadString(url);
             return JsonConvert.DeserializeObject<T>(response);
+        }
+        
+        private string RepositoryName
+        {
+            get { return string.Format("{0}/{1}", Account, Repo); }
+        }
+
+        private T GetValue<T>(string key, Func<T> defaultValue)
+        {
+            return _storage.ContainsKey(key) ? _storage.Get<T>(key) : defaultValue();
         }
     }
 }
